+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 CHANGELOG
 =========
 
+* it's not possible any more to show tests that don't belong to the course
+  being shown
 * pass the right context when accessing a course from the course list
 * don't break when the Mooc.IP overview navigation item is not available due
   to some misconfiguration